+++ packages/svg-react-transformer/lib/templates/fancy.js
@@ -24,22 +24,31 @@ function fancy(data) {
     const React = require('react');
     class ${data.name} extends React.PureComponent {
       render() {
-        const containerStyle = this.props.containerStyle || {};
+        const containerStyle = {};
+        Object.keys(this.props.containerStyle || {}).forEach(key => {
+          containerStyle[key] = this.props.containerStyle[key];
+        });
         if (!containerStyle.position || containerStyle.position === 'static') {
           containerStyle.position = 'relative';
         }
         containerStyle.paddingBottom = '${ratio}%';
-        const svgStyle = this.props.svgStyle || {};
+
+        const svgStyle = {};
+        Object.keys(this.props.svgStyle || {}).forEach(key => {
+          svgStyle[key] = this.props.svgStyle[key];
+        });
         svgStyle.position = 'absolute';
         svgStyle.overflow = 'hidden';
         svgStyle.top = 0;
         svgStyle.left = 0;
         svgStyle.width = '100%';
+
         const text = !this.props.alt ? null : (
           <div style={{ position: 'absolute', left: -9999 }}>
             {this.props.alt}
           </div>
         );
+
         return (
           <div
             style={containerStyle}